Some renaming
* ObjectsFilterModel > ReversingRecursiveFilterModel
* ImmutableRoleModel > ImmutableMapObjectProxyModel
* Immutable map objects model is now even more immutable
