+++ src/tiled/tiled.qbs
@@ -347,7 +347,6 @@ QtGuiApplication {
         "objectselectionitem.h",
         "objectselectiontool.cpp",
         "objectselectiontool.h",
-        "objectsfiltermodel.h",
         "objectsview.cpp",
         "objectsview.h",
         "objecttemplatemodel.cpp",
@@ -410,6 +409,7 @@ QtGuiApplication {
         "resizetilelayer.h",
         "reversingproxymodel.cpp",
         "reversingproxymodel.h",
+        "reversingrecursivefiltermodel.h",
         "rotatemapobject.cpp",
         "rotatemapobject.h",
         "scriptedaction.cpp",
